# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "description": "Research Support Tool API Documentation",
    "version": "1.0.0",
    "title": "Research Support Tool API",
    "contact": {
      "email": "guillem.miro@students.salle.url.edu,faozi.bouybaouene@students.salle.url.edu",
      "name": "Guillem Miró Sierra & Faozi Bouybaouene"
    },
    "license": {
      "name": "GNU GPL v3 or later",
      "url": "http://www.gnu.org/copyleft/gpl.html"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "landing",
      "description": "API Calls per la gestió de la pàgina d'inici"
    },
    {
      "name": "journals",
      "description": "API Calls per a la gestió dels journals"
    },
    {
      "name": "specialIssues",
      "description": "API Calls per a la gestió dels special issues"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/journals/": {
      "get": {
        "tags": [
          "journals"
        ],
        "description": "Retorna tots els journals",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "journals"
        ],
        "description": "Afegir un nou journal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "journal",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Journal"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/journals/{title}": {
      "get": {
        "tags": [
          "journals"
        ],
        "description": "Retorna tots els journals amb el títol passat per paràmetre",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "journals"
        ],
        "description": "Esborra el journal amb el títol passat per paràmetre",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/journals/search/{title}": {
      "get": {
        "tags": [
          "journals"
        ],
        "description": "Retorna els journals amb el títol passat per paràmetre",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/journals/issn/{issn}": {
      "get": {
        "tags": [
          "journals"
        ],
        "description": "Retorna el journal amb el issn passat per paràmetre",
        "parameters": [
          {
            "name": "issn",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/journals/tag/{tag}": {
      "get": {
        "tags": [
          "journals"
        ],
        "description": "Retorna tots els journals amb els tags corresponents al paràmetre passat",
        "parameters": [
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/journals/tags": {
      "post": {
        "tags": [
          "journals"
        ],
        "description": "Retorna els journals amb els tags passats per paràmetre en el body",
        "parameters": [
          {
            "name": "tags",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/specialIssues/": {
      "get": {
        "tags": [
          "specialIssues"
        ],
        "description": "Retorna tots els special issues",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "specialIssues"
        ],
        "description": "Afegir un nou special issue",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "specialIssue",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SpecialIssue"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/specialIssues/{title}": {
      "get": {
        "tags": [
          "specialIssues"
        ],
        "description": "Retorna tots els special issues amb el títol passat per paràmetre",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "specialIssues"
        ],
        "description": "Esborra el special issue amb el títol passat per paràmetre",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/specialIssues/search/{title}": {
      "get": {
        "tags": [
          "specialIssues"
        ],
        "description": "Retorna el special issue amb el titol passat per paràmetre",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/specialIssues/tag/{tag}": {
      "get": {
        "tags": [
          "specialIssues"
        ],
        "description": "Retorna els special issues amb el tag passat per paràmetre",
        "parameters": [
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/specialIssues/tags": {
      "post": {
        "tags": [
          "specialIssues"
        ],
        "description": "Retorna els special issues amb els tags passats per paràmetre en el body",
        "parameters": [
          {
            "name": "tags",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/landing/login": {
      "get": {
        "tags": [
          "landing"
        ],
        "description": "Retorna un token JWT per a l'usuari amb el usuari i la contrasenya passats per paràmetre",
        "parameters": [
          {
            "name": "username",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/landing/getNumbersByDocumentType": {
      "get": {
        "tags": [
          "landing"
        ],
        "description": "Retorna el nombre de documents per a cada tipus de document",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/landing/getDocumentsByDocumentType": {
      "get": {
        "tags": [
          "landing"
        ],
        "description": "Retorna tots els documents per a cada tipus de document",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}